+++ apps/website/src/routes/docs/headless/tabs/index.mdx
@@ -85,6 +85,27 @@ You can pass a custom `onClick$` handler.
 
 <Showcase name="on-click" />
 
+## Creating reusable Tabs
+
+In order to remove the need for a linking `value` prop for each Tab and TabPanel pair, we have chosen to create the Tabs component as an [inline component](https://qwik.builder.io/docs/components/overview/#inline-components).
+
+By consequence, the Tabs component needs to be aware of its children. If you want to create your custom reusable TabList/Tab/TabPanel components, you will have to pass them to the Tabs component through its `TabList`, `Tab`, and `TabPanel` props:
+
+```tsx
+const CustomTabs = (props: TabsProps) => (
+  <Tabs {...props} TabList={CustomTabList} Tab={CustomTab} TabPanel={CustomTabPanel} />
+);
+```
+
+<br />
+
+<Note status="warning">
+  If you don't do the above, the Tabs component cannot know if your CustomTab component is
+  a Tab component.
+</Note>
+
+<Showcase name="reusable" />
+
 ## Accessibility
 
 ### Keyboard interaction
